using shared jest config. Fixed mapping tests
