+++ database-bridge-lambda/fileTransformer.js
