+++ database-bridge-lambda/jest.config.js
